ADD: argv() option "prefix_dashfiles"
* "prefix_dashfiles" option
  - causes all files == "-" to be converted an equivalent prefixed version (eg, "./-")
  - on by default

.# Discussion

In general, a "-" used as a "filename" is interpreted to mean standard input (aka,
STDIN). So, any globbed argument producing a file named "-" would be confused as STDIN.

To avoid this, when globbing, if a file with a globbed path equal to "-" is found, it
is renamed to the equivalent "./-".

Non-globbed arguments are left alone.